click handler defined on sidepanel
